# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Olena Studio",
    "version": "1.0.9",
    "description": "home project",
    "contact": {
      "name": "Maria Skydan",
      "email": "mariaskydan@gmail.com",
      "url": "https://github.com/maskaa22?page=1&tab=repositories"
    }
  },
  "parameters": {
...
    "userUpdateDate": {
      "name": "updateDate",
      "in": "body",
      "schema": {
        "type": "object",
        "default": {
          "_id": "64159105f74553185be6c182",
          "name": "",
          "surname": "",
          "email": "",
          "phone": "0667845588",
          "oldPassword": "",
          "number": "",
          "numberToo": "",
          "nameSity": "",
          "nameDepartment": ""
        }
      }
    }
  },
  "paths": {
...
    "/users": {
      "get": {
        "tags": [
          "users"
        ],
        "operationId": "getAllUsers",
        "summary": "Перегляд списку зареєстрованих користувачів",
        "description": "Тільки авторизований адміністратор може отримати список всіх користувачів",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "type": "string",
            "required": true,
            "default": "Bearer",
            "description": "Після слова Bearer, через пробіл вставте свій access_token"
          }
        ],
        "responses": {
          "200": {
            "description": "Список користувачів",
            "schema": {
              "$ref": "#/definitions/userSchema"
            }
          },
          "400": {
            "description": "Error: Bad Request",
            "schema": {
              "$ref": "#/definitions/errorSchemaNotUser"
            }
          },
          "404": {
            "description": "Error: Not Found",
            "schema": {
              "$ref": "#/definitions/errorSchemaNotFoundUsers"
            }
          }
        }
      },
      "patch": {
        "tags": [
          "users"
        ],
        "operationId": "updateAllDate",
        "summary": "Внесення змін в усі дані",
        "description": "Обов'язкове заповнення поля '_id'. Для зміни пароля введіть старий пароль(oldPassword), та новий два рази(number, numberToo). Для будь яких інших змін, використовується тільки одне поле. Для прикладу змінимо номер телефону на новий.",
        "parameters": [
          {
            "$ref": "#/parameters/userUpdateDate"
          }
        ],
        "responses": {
          "200": {
            "description": "Дані збережені",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "example": "Дані збережені"
                }
              }
            }
          },
          "404": {
            "description": "Error: Not Found",
            "schema": {
              "$ref": "#/definitions/errorSchemaNotFound"
            }
          },
          "409": {
            "description": "Error: Conflict",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "example": "Для зміни пароля заповніть поля"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "users"
        ],
        "operationId": "updateFoto",
        "summary": "Зміна фотокартки",
        "description": "Введіть ідентифікатор користувача та завантажте фото.",
        "parameters": [
          {
            "name": "_id",
            "in": "formData",
            "type": "string",
            "default": "64159105f74553185be6c182",
            "required": true
          },
          {
            "name": "foto",
            "in": "formData",
            "type": "file",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Користувач",
            "schema": {
              "$ref": "#/definitions/userSchemaForUpdate"
            }
          }
        }
      }
    },
...
    "/users/send": {
      "post": {
        "tags": [
          "users"
        ],
        "operationId": "sendUser",
        "summary": "Написання листа користувачеві.",
        "description": "Текст повідомлення - 'text', пошта на яку слати листа - 'email' та його тема - 'topic'.",
        "parameters": [
          {
            "name": "sendUser",
            "in": "body",
            "schema": {
              "type": "object",
              "default": {
                "text": "Вітаємо. На передодні свят запрошуємо до нас на чарівні зачіски.",
                "email": "sokolavanila22@gmail.com",
                "topic": "З найкращими побажаннями"
              }
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/emailAcceptedSchema"
            }
          },
          "400": {
            "description": "Error: Bad Request",
            "schema": {
              "$ref": "#/definitions/errorSchemaNotFound"
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "errorSchemaNotFound": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "description": "Текст повідомлення",
          "example": "Заповніть поля"
        }
      }
    },
...
    "errorSchemaNotFoundUsers": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "description": "Текст повідомлення",
          "example": "Користувачі відсутні"
        }
      }
    },
    "errorSchemaNotUser": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "description": "Текст повідомлення",
          "example": "Тільки для адміністратора"
        }
      }
    },
...
    "emailAcceptedSchema": {
      "type": "array",
      "items": {
        "properties": {
          "accepted": {
            "type": "string",
            "description": "Емеіл відправника",
            "example": "<ead91dad-0044-7c4d-444d-61720a56d877@example.com>"
          },
          "rejected": {
            "type": "array",
            "example": []
          },
          "envelopeTime": {
            "type": "number",
            "example": 315
          },
          "messageTime": {
            "type": "number",
            "example": 632
          },
          "messageSize": {
            "type": "number",
            "example": 575
          },
          "response": {
            "type": "string",
            "example": "250 2.0.0 OK  1680619585 dn25-20020a05640222f900b004be11e97ca2sm5893504edb.90 - gsmtp"
          },
          "envelope": {
            "type": "object",
            "properties": {
              "from": {
                "description": "Від кого лист",
                "type": "string",
                "example": "g@example.com"
              },
              "to": {
                "description": "Кому лист(емеіл)",
                "type": "string",
                "example": "lyi@gmail.com"
              }
            }
          },
          "messageId": {
            "type": "string",
            "example": "<ead91dad-0044-7c4d-444d-61720a56d877@example.com>"
          }
        }
      }
    },
...
    "userSchemaForUpdate": {
      "properties": {
        "_id": {
          "description": "Mongo ідентифікатор",
          "example": "555df2q5dcf"
        },
        "name": {
          "description": "Ім'я користувача",
          "type": "string",
          "example": "Луі"
        },
        "email": {
          "description": "Емеіл користувача",
          "type": "string",
          "example": "sds@example.com"
        },
        "role": {
          "description": "Роль користувача (user/admin)",
          "type": "string",
          "example": "user"
        },
        "foto": {
          "description": "Назва фотокартки",
          "type": "string",
          "example": "2dcsdvs-16sdfvdd-4b6d-925b-e3sgddfh6d7.jpg"
        },
        "is_active": {
          "description": "Підтвердження емейлу",
          "type": "boolean",
          "example": false
        },
        "createdAt": {
          "description": "Дата створення користувача",
          "type": "string",
          "example": "2023-12-13T10:23:01.414Z"
        },
        "updatedAt": {
          "description": "Дата внесення змін",
          "type": "string",
          "example": "2023-12-13T10:24:38.892Z"
        },
        "__v": {
          "description": "Версійність в Mongo",
          "type": "integer",
          "example": 0
        },
        "surname": {
          "description": "Прізвище користувача",
          "type": "string",
          "example": "Сальваторе"
        },
        "phone": {
          "description": "Номер телефону",
          "type": "string",
          "example": "0987793157"
        },
        "nameSity": {
          "description": "Місто одержання посилки для Нової пошти",
          "type": "string",
          "example": "Вінниця"
        },
        "nameDepartment": {
          "description": "Номер відділення Нової пошти",
          "type": "string",
          "example": "25"
        }
      }
    },
    "acessTokenSchema": {
      "type": "string",
      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjY0MTU5MTA1Zjc0NsdgdhSGSEGjkk4MDQxODgzNCwiZXhwIjoxNjgwNDE5NzM0fQ.rCvXJn4CfpjWp6lOgB7rk-M02IdHKsXVEdrRr9KupJI"
    },
    "refreshTokenSchema": {
      "type": "string",
      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjY0MTU5MTA1Zjc0NsdgdhSGSEGjkk4MDQxODgzNCwiZXhwIjoxNjgwNDE5NzM0fQ.rCvXJn4CfpjWp6lOgB7rk-M02IdHKsXVEdrRr9KupJI"
    },
    "userSchema": {
      "type": "array",
      "items": {
        "properties": {
          "user": {
            "type": "object",
            "$ref": "#/definitions/userSchemaForUpdate"
          },
          "access_token": {
            "type": "object",
            "$ref": "#/definitions/acessTokenSchema"
          },
          "refresh_token": {
            "type": "object",
            "$ref": "#/definitions/refreshTokenSchema"
          }
        }
      }
    },
...
  }
}